= Minor =  * Fix body re  * Split the button image creation into a function for ease of use.  * Fixes to the panel system  * Added indexOf convenience method to Array objects    {{{ 3 == ['a','b','c'].indexOf('c'); }}}  * Fix to the font dropdown updating (missing break)  * Focus the popup editor window.  Not sure why I fixed that, because popupeditor is going to be removed totaly (in favour of FullScreen).
= Major =
 * Add an auto-link ability for Gecko.  This is intended to work the same way as IE, 
for example if you type www.example.com then IE automatically makes that into a link, it 
does the same for email addresses.  It's slightly better than IE's method though as this 
allows you to easily undo the "damage" if you didn't want the link created, just hit 
CTRL-Z (undo) or ESC (escape) immediatly after the link is created and it gets unlinked.  
Note: this adds a lot of code.  One day the IE and Gecko specific code should be split out 
so it's not loaded if it's not needed.
 * New config option "mozParahandler", this should be set to built-in, dirty or best
   * built-in doesn't alter mozilla's paragraph handling at all, so this probably means 
hitting enter will cause a <br/> to be inserted.
   * dirty will use the "quick and dirty" fix which will make mozilla put in paragraphs 
instead of breaks in most cases.  But it's far from perfect.
   * best will use "hipikat"'s EnterParagraphs plugin to provide the fix for mozilla 
(loading EnterParagraphs is not necessary it is done automatically if "best" is selected).  
The EnterParagraphs plugin will give the best results, but it will be slower about it.




git-svn-id: http://svn.xinha.org/trunk@9 bdcdd039-d5ef-0310-875c-f7949af66bfe