Fixes to problem shown in ticket:237 that was introduced in changeset:69
This change puts it back so that only one editor is active at any given 
time, and that editors must initially be activated by a mouse-down on 
the area by the user (in both gecko and IE now).

This may fix some other tickets with similar types of problem 
(insertHTML placing text outside of the editor).  The problem mostly 
came down to focusEditor() not doing work when it should have because of 
a misplaced bang, but some other issues needed cleaning up too.



git-svn-id: http://svn.xinha.org/trunk@160 bdcdd039-d5ef-0310-875c-f7949af66bfe