There are 4 things in this commit...
1. Resizing fixes.
2. New "flowable" toolbar
3. Proof of concept of skins.
4. Testbed file.

=== Resizing Fixes ===
Complete rewrite of the framing that the editor is built from, 
previously it was done using nested and positioned divs, which was very 
troublesome to get right.  It also suffered from problems when resizing 
the editor.

I have changed it to now use a table structure and have fixed the 
resizing problems.  This also means we can now use any CSS measurement 
for setting width and height (even percentages).  Resizing is handled 
very well, and generally it's just better all around.

The fullscreen plugin has been adjusted to use the new methods, which 
makes it's job much easier too.  This won't solve the problem with 
fullscreen when the editor has been placed inside a positioned element 
however.

=== Flowable Toolbar ===
A new config option has been added which allows the toolbar sections 
(between separators) to be treated as "possible-line-break-here", 
meaning that if there is not enough room on the line, the line will 
break to put the section on the next line.  Newlines are treated the 
same way.  Of course the toolbar will re-flow when you resize the 
window.

This has been set as the default (and the toolbar has been adjusted to 
better suit it) because the default toolbars can get pretty long as you 
add plugins.  Having the flowing toolbars alleviates this problem and 
allows a very small editor window :)

=== Skins ===
Ok, well, it's just a CSS file and a couple of images, but it shows the 
sort of thing we can already do.  I felt that this is something that 
people would want to see.  What I'd like in the future is to be able to 
have a config option per editor to select what skin it gets, but this 
will necessitate some extra work so we can have different styles apply 
to each editor on a multi-editor page.

The other thing that would be nice is to be able to style the select 
lists in the toolbar, which are pretty ugly.

I've set the "blue-look" skin to be used in the full example.

=== Testbed ===
This is just a simple html file using Xinha without all the extra stuff 
in the full_example.  You can use it to test your plugins etc, or users 
could take it as a boilerplate file.



git-svn-id: http://svn.xinha.org/trunk@166 bdcdd039-d5ef-0310-875c-f7949af66bfe