Ticket #298 Undo and Redo are completely broken in full page mode [[BR]]
this._doc.setFullHTML(html) is not a function [[BR]]
this.setFullHTML(html) is to set but thats not all! [[BR]]

With a little Testfile it works but with the Example it doesn't works! [[BR]]
HTMLArea.RE_body    = /<body[^>]*>((.|\n)*?)<\/body>/i; [[BR]]
is good but [[BR]]
HTMLArea.RE_body    = /<body[^>]*>((.|\n|\r|\t)*?)<\/body>/i; [[BR]]
is better! [[BR]]

git-svn-id: http://svn.xinha.org/trunk@258 bdcdd039-d5ef-0310-875c-f7949af66bfe