ticket:1518
Changeset requires that if a non default key location is used in xinha_pass_to_php_backend, that the same key location must also be passed as a parameter to xinha_read_passed_data.
Potentially this whole non-default thing could be stripped out as it's perhaps not that useful anyway, but seems reasonable to minimise the change in this manner.



git-svn-id: http://svn.xinha.org/trunk@1251 bdcdd039-d5ef-0310-875c-f7949af66bfe