Updates to http://www.github.com/sleemanj/mootools-filemanager from GerHoblett   http://www.github.com/GerHoblett/ Changes to said updates by gogo (sleemanj @ github)
Modifications to MootoolsFileManager to work with those updates, some courtesy of GerHoblett, some sleemanj

GerHoblett provided a large diff which accomplished the goal in a quite different way.  It has merit, however I have opted for a less-affecting path in so far as Xinha's "way" is concerned, namely, not splitting the config for a single plugin into several calls to backend config passing functions which seemed a little cumbersome.

Instead I take the option of using POST to send backend data around, at the minor expense of an extra round trip when displaying thumbnails (for each one).  This could be reduced by checking for thumbnail existence and returning the thumbnail name directly in "onView" rather than the backend request to generate said thumbnail.

Still to do, is to make the preview pane thumbnail also work.

  


git-svn-id: http://svn.xinha.org/branches/MootoolsFileManager-Update@1302 bdcdd039-d5ef-0310-875c-f7949af66bfe